+++ src/main/java/nl/eveoh/mytimetable/blackboard/service/ConfigService.java
@@ -76,6 +76,11 @@ public class ConfigService extends HttpServlet {
             messages.put("applicationTarget", "Target should be set.");
         }
 
+        String applicationUriDescriptionOverride = request.getParameter("applicationUriDescriptionOverride");
+        if (StringUtils.isBlank(applicationUriDescriptionOverride)) {
+            applicationUriDescriptionOverride = null;
+        }
+
         int numberOfEvents = -1;
         try {
             numberOfEvents = Integer.parseInt(request.getParameter("numberOfEvents"));
@@ -170,9 +175,10 @@ public class ConfigService extends HttpServlet {
             // Save the configuration.
 
             try {
-                saveConfig(applicationUri, selectedTarget, numberOfEvents, defaultNumberOfEvents, showActivityType,
-                        apiEndpointUris, apiKey, apiSslCnCheck, apiConnectTimeout, apiSocketTimeout, apiMaxConnections,
-                        usernameDomainPrefix, usernamePostfix, customCss, timetableTypes, unknownLocationDescription);
+                saveConfig(applicationUri, selectedTarget, applicationUriDescriptionOverride, numberOfEvents,
+                        defaultNumberOfEvents, showActivityType, apiEndpointUris, apiKey, apiSslCnCheck,
+                        apiConnectTimeout, apiSocketTimeout, apiMaxConnections, usernameDomainPrefix, usernamePostfix,
+                        customCss, timetableTypes, unknownLocationDescription);
             } catch (ConfigurationPersistenceException e) {
                 log.error("Something went wrong with saving the preferences", e);
 
@@ -186,11 +192,12 @@ public class ConfigService extends HttpServlet {
         }
     }
 
-    private void saveConfig(String applicationUri, String selectedTarget, int maxNumberOfEvents,
-                            int defaultNumberOfEvents, boolean showActivityType, ArrayList<String> apiEndpointUris,
-                            String apiKey, boolean apiSslCnCheck, int apiConnectTimeout, int apiSocketTimeout,
-                            int apiMaxConnections, String usernameDomainPrefix, String usernamePostfix,
-                            String customCss, String timetableTypesStr, String unknownLocationDescription) {
+    private void saveConfig(String applicationUri, String selectedTarget, String applicationUriDescriptionOverride,
+                            int maxNumberOfEvents, int defaultNumberOfEvents, boolean showActivityType,
+                            ArrayList<String> apiEndpointUris, String apiKey, boolean apiSslCnCheck,
+                            int apiConnectTimeout, int apiSocketTimeout, int apiMaxConnections,
+                            String usernameDomainPrefix, String usernamePostfix, String customCss,
+                            String timetableTypesStr, String unknownLocationDescription) {
         try {
             WidgetConfiguration configuration = ConfigUtil.loadConfig();
 
@@ -217,6 +224,8 @@ public class ConfigService extends HttpServlet {
                 configuration.getTimetableTypes().clear();
                 configuration.getTimetableTypes().addAll(timetableTypes);
             }
+
+            configuration.setApplicationUriDescriptionOverride(applicationUriDescriptionOverride);
 
             ConfigUtil.saveConfig(configuration);